Introduce abstract base class for Apdb.
Apdb class is now an abstract interface, SQL-based implementation is in
ApdbSql class. Configuration classes are also a hierarchy with common
parameters in ApdbConfig class and SQL-specific fields in ApdbSqlConfig.
Configuration should work better with a ConfigurableField.

Also changed logic for `isolation_level` for sqlite to avoid the need to
override default in every case.
